Module Tjr_lru_cache

A multithreaded LRU cache (part of ImpFS; safe to open)

NOTE Mt_ops may have field names that clash with those elsewhere

include Tjr_lru_cache__.Im_intf.Entry_type

type 'v entry =

| Insert of {

value : 'v;

dirty : bool;

}

| Delete of {

dirty : bool;

}

| Lower of 'v option

Cache map entries; values in the map are tagged with a last-accessed time (via underlying Lru impl) and a dirty flag

Entries in the cache for key k:

(Insert v (dirty=true/false)), this occurs on insert

(Delete (dirty=true/false))

(Lower vopt), this occurs when we check the lower layer for a non-existing entry in cache; if we find a value, we insert Lower (Some v), else Lower None; in either case, there is no need to do anything further (ie the entry is not dirty)

(No entry), for a key that hasn't been seen before

type persist_mode = Tjr_lru_cache__.Mt_intf.persist_mode =

| Persist_later

| Persist_now

include Tjr_lru_cache__.Mt_intf.Mt_ops

type ('k, 'v, 't) mt_ops = {

mt_find : 'k -> ('v option, 't) Tjr_monad.m;

mt_insert : Tjr_lru_cache__.Mt_intf.persist_mode -> 'k -> 'v -> (unit, 't) Tjr_monad.m;

mt_delete : Tjr_lru_cache__.Mt_intf.persist_mode -> 'k -> (unit, 't) Tjr_monad.m;

mt_sync_key : 'k -> (unit, 't) Tjr_monad.m;

mt_sync_all_keys : unit -> (unit, 't) Tjr_monad.m;

}

include Tjr_lru_cache__.Mt_intf.Lru_msg_type

type ('k, 'v, 't) lru_msg =

| Insert of 'k * 'v * unit -> (unit, 't) Tjr_monad.m

| Delete of 'k * unit -> (unit, 't) Tjr_monad.m

| Find of 'k * 'v option -> (unit, 't) Tjr_monad.m

| Evictees of ('k * 'v Tjr_lru_cache__.Im_intf.entry) list

The type of messages that we send to the lower level. NOTE that the callbacks are needed to implement the "persist now" mode, i.e., to inform the caller when the operation has committed to disk.

include Tjr_lru_cache__.Mt_intf.Lru_factory

type ('k, 'v, 'lru, 't) lru_factory = < max_size : int; evict_count : int; empty : 'lru; make : with_lru:('lru, 't) Tjr_monad.with_state -> to_lower:(('k, 'v, 't) Tjr_lru_cache__.Mt_intf.Lru_msg_type.lru_msg -> (unit, 't) Tjr_monad.m) -> ('k, 'v, 't) Tjr_lru_cache__.Mt_intf.mt_ops; make_with_ref : to_lower:(('k, 'v, 't) Tjr_lru_cache__.Mt_intf.Lru_msg_type.lru_msg -> (unit, 't) Tjr_monad.m) -> < ops : ('k, 'v, 't) Tjr_lru_cache__.Mt_intf.mt_ops; lru_ref : 'lru Stdlib.ref; with_lru : ('lru, 't) Tjr_monad.with_state; >; >

This is the eventual interface provided by Make. NOTE this type is unreadable in odoc

type ('k,'v,'lru,'t) lru_factory = <
  size: int;
  evict_count: int;
  empty: 'lru;
  make: 
    with_lru:('lru,'t) with_state ->
    to_lower:(('k,'v,'t)lru_msg -> (unit,'t)m) -> 
    ('k,'v,'t)mt_ops;
  make_with_ref:
    to_lower:(('k,'v,'t)lru_msg -> (unit,'t)m) -> 
    < ops:('k,'v,'t)mt_ops;
      lru_ref:'lru ref;
      with_lru: ('lru,'t) with_state >
>

module Make_1 : sig ... end

Construct the multithreaded lru (functor version)

module Make_2 : sig ... end

Construct the multithreaded lru; like Make_1, but targetting the lru factory

module Make_3 : sig ... end

Construct the multithreaded LRU; like Make_2, but specialized to Lwt

module Make = Make_3