Module Tjr_lru_cache.Lru_in_mem

A single-threaded in-memory LRU cache, as a prelude to the real thing.

We assume all operations complete quickly from the cache. Operations that need to go to disk are modelled using explicit continuation passing.

We allow values to be tagged with extra info (this will be used to record whether the value needs to be flushed to the lower level). ?FIXME

The difficulty is how to extend this to deal with disk-backed eviction.

NOTE that the operations occur not in a monad - instead, explicit state passing is used, and we return the possible evictees; this is a prelude to the real version where we need to do something with the evictees

module Make_lru_fc : functor (S : sig ... end) -> sig ... end

module Internal : functor (S : sig ... end) -> sig ... end

val make_lru_in_mem_ops : ('k, 'v, 'lru) Im_intf.lru_fc -> ('k, 'v, 'lru) Im_intf.lim_ops

val make_init_lim_state : max_size:int -> evict_count:int -> lru_fc:('k, 'v, 'lru) Im_intf.lru_fc -> ('k, 'v, 'lru) Im_intf.lim_state

module Int_int : sig ... end

Don't open this directly; use Int_int.ops etc