Module Lru_in_mem.Internal

Parameters

S : sig ... end

Signature

module L : sig ... end

type lru = L.t

val find_in_cache : S.k -> (S.k, S.v, lru) Im_intf.lim_state -> L.v option * (S.k, S.v, lru) Im_intf.lim_state

Attempt to locate key in cache. FIXME inline

exception E_

An exception, for quick abort. FIXME remove

val make_cached_map : unit -> (S.k -> (S.k, S.v, lru) Im_intf.lim_state -> (L.v, vopt_from_lower:S.v option -> lim_state:(S.k, S.v, lru) Im_intf.lim_state -> S.v option * (S.k, S.v, lru) Im_intf.evictees_x_lim_state) Im_intf.maybe_in_cache) * (L.k -> S.v -> ('a, 'b, L.t) Im_intf.lim_state -> (S.k, S.v, lru) Im_intf.evictees_x_lim_state) * (L.k -> ('c, 'd, L.t) Im_intf.lim_state -> (S.k, S.v, lru) Im_intf.evictees_x_lim_state) * (L.k -> (S.k, S.v, lru) Im_intf.lim_state -> (L.v * ('e, 'f, L.t) Im_intf.lim_state, unit) Im_intf.maybe_in_cache)

Construct the cached map, using existing map find operation.

NOTE the idea for find is that we execute a quick step to handle the case that there is an entry in the cache. If there isn't, we make a slow call to the lower layer, and when the result arrives we make another update to the cache state with the then current state of the cache. To avoid the risk of stale results being returned from lower, we have to tag lower results with some kind of monotonically-increasing index.

val make_lru_in_mem : unit -> (L.k, S.v, lru) Im_intf.lim_ops

Package up the operations in a record

val make_init_lim_state : max_size:int -> evict_count:int -> ('a, 'b, L.t) Im_intf.lim_state