Module Tjr_lru_cache__.Im_intf

Cache entries

module Entry : sig ... end

include Entry

type 'v entry =

| Insert of {

value : 'v;

dirty : bool;

}

| Delete of {

dirty : bool;

}

| Lower of 'v option

Cache map entries; values in the map are tagged with a last-accessed time (via underlying Lru impl) and a dirty flag

Entries in the cache for key k:

(Insert v (dirty=true/false)), this occurs on insert

(Delete (dirty=true/false))

(Lower vopt), this occurs when we check the lower layer for a non-existing entry in cache; if we find a value, we insert Lower (Some v), else Lower None; in either case, there is no need to do anything further (ie the entry is not dirty)

(No entry), for a key that hasn't been seen before

val is_Lower : 'a entry -> bool

val mark_clean : 'a entry -> 'a entry

val entry_is_dirty : 'a entry -> bool

Lru cache state (in memory)

module Lru_state_im : sig ... end

The LRU in-memory cache state consists of:

include Lru_state_im

type 'lru lru_state_im = {

max_size : int;

evict_count : int;

lru_state : 'lru;

compare_lru : 'lru -> 'lru -> int;

}

module Lru_ops_im : sig ... end

include Lru_ops_im

type ('a, 'b) maybe_in_cache =

| In_cache of 'a

| Not_in_cache of 'b

type ('k, 'v, 'lru) evictees_x_state = {

evictees : ('k * 'v entry) list option;

lru_state_im : 'lru lru_state_im;

}

type ('k, 'v, 'lru) lru_ops_im = {

find : 'k -> 'lru lru_state_im -> ('v entry, vopt_from_lower:'v option -> lru_state_im:'lru lru_state_im -> 'v option * ('k, 'v, 'lru) evictees_x_state) maybe_in_cache;

insert : 'k -> 'v -> 'lru lru_state_im -> ('k, 'v, 'lru) evictees_x_state;

delete : 'k -> 'lru lru_state_im -> ('k, 'v, 'lru) evictees_x_state;

sync_key : 'k -> 'lru lru_state_im -> ('v entry * 'lru lru_state_im, unit) maybe_in_cache;

}

This type is what is returned by the make_lru_in_mem function.

NOTE that sync_key performs only the in-mem updates (ie clearing the dirty flag). If you want to flush to disk, you have to do something else (see Lru_multithreaded).

NOTE find returns a "maybe_in_cache", which consists of a callback function which tries to retrieve from below, and then on return updates the then-current lim_state (which may already have the entry...)

NOTE the interface tries to favour pure state passing; find is the only routine which may need to call to disk