Module Im_intf.Lim_ops

This type is what is returned by the make_lru_in_mem function.

NOTE that sync_key performs only the in-mem updates (ie clearing the dirty flag). If you want to flush to disk, you have to do something else (see Lru_multithreaded).

NOTE find returns a "maybe_in_cache", which consists of a function which tries to retrieve from below, and then on return updates the then-current lim_state (which may already have the entry...)

NOTE the interface tries to favour pure state passing; find is the only routine which may need to call to disk

NOTE a struct so that we can scope record fields as Lim_ops.{ ... }

type ('k, 'v, 'lru) lim_ops = {

find : 'k -> ('k, 'v, 'lru) lim_state -> ('v entry, vopt_from_lower:'v option -> lim_state:('k, 'v, 'lru) lim_state -> 'v option * ('k, 'v, 'lru) evictees_x_lim_state) maybe_in_cache;

insert : 'k -> 'v -> ('k, 'v, 'lru) lim_state -> ('k, 'v, 'lru) evictees_x_lim_state;

delete : 'k -> ('k, 'v, 'lru) lim_state -> ('k, 'v, 'lru) evictees_x_lim_state;

sync_key : 'k -> ('k, 'v, 'lru) lim_state -> ('v entry * ('k, 'v, 'lru) lim_state, unit) maybe_in_cache;

}