Module Im_intf.Lru_ops_im

Lru ops (in memory)

type ('a, 'b) maybe_in_cache =

| In_cache of 'a

| Not_in_cache of 'b

type ('k, 'v, 'lru) evictees_x_state = {

evictees : ('k * 'v entry) list option;

lru_state_im : 'lru lru_state_im;

}

type ('k, 'v, 'lru) lru_ops_im = {

find : 'k -> 'lru lru_state_im -> ('v entry, vopt_from_lower:'v option -> lru_state_im:'lru lru_state_im -> 'v option * ('k, 'v, 'lru) evictees_x_state) maybe_in_cache;

insert : 'k -> 'v -> 'lru lru_state_im -> ('k, 'v, 'lru) evictees_x_state;

delete : 'k -> 'lru lru_state_im -> ('k, 'v, 'lru) evictees_x_state;

sync_key : 'k -> 'lru lru_state_im -> ('v entry * 'lru lru_state_im, unit) maybe_in_cache;

}

This type is what is returned by the make_lru_in_mem function.

NOTE that sync_key performs only the in-mem updates (ie clearing the dirty flag). If you want to flush to disk, you have to do something else (see Lru_multithreaded).

NOTE find returns a "maybe_in_cache", which consists of a callback function which tries to retrieve from below, and then on return updates the then-current lim_state (which may already have the entry...)

NOTE the interface tries to favour pure state passing; find is the only routine which may need to call to disk