Module Make_v1.Make_cached_map

Construct the cached map, using existing map find operation.

NOTE the idea for find is that we execute a quick step to handle the case that there is an entry in the cache. If there isn't, we make a slow call to the lower layer, and when the result arrives we make another update to the cache state with the then current state of the cache. To avoid the risk of stale results being returned from lower, we have to tag lower results with some kind of monotonically-increasing index.

val get_evictees : lru Tjr_lru_cache.Im_intf.lru_state_im -> (S.k, S.v, lru) Tjr_lru_cache.Im_intf.evictees_x_state

val find : S.k -> lru Tjr_lru_cache.Im_intf.lru_state_im -> [> `In of S.v option | `Not_in ]

val perform : L.k -> L.v -> lru Tjr_lru_cache.Im_intf.lru_state_im -> (S.k, S.v, lru) Tjr_lru_cache.Im_intf.evictees_x_state

val insert : L.k -> S.v -> lru Tjr_lru_cache.Im_intf.lru_state_im -> (S.k, S.v, lru) Tjr_lru_cache.Im_intf.evictees_x_state

val delete : L.k -> lru Tjr_lru_cache.Im_intf.lru_state_im -> (S.k, S.v, lru) Tjr_lru_cache.Im_intf.evictees_x_state

val update_from_lower : L.k -> S.v option -> lru Tjr_lru_cache.Im_intf.lru_state_im -> (S.k, S.v, lru) Tjr_lru_cache.Im_intf.evictees_x_state

val sync_key : S.k -> lru Tjr_lru_cache.Im_intf.lru_state_im -> ((S.k, S.v) Tjr_fs_shared.kvop * lru Tjr_lru_cache.Im_intf.lru_state_im) option

val sync_all_keys : lru Tjr_lru_cache.Im_intf.lru_state_im -> (L.k, S.v, lru) Tjr_lru_cache.Im_intf.evictees_x_state