Module Mt_intf.Mt_state_type

type ('k, 'v, 'lru, 't_map, 't) mt_state = {

lim_state : ('k, 'v, 'lru) Tjr_lru_cache__.Im_intf.lim_state;

blocked_threads : 't_map;

blocked_threads_ops : ('k, ('v, 't) Threading_types.blocked_thread list, 't_map) Tjr_lib.Tjr_map.map_ops;

}

The state consists of:

the cache state

a map from k to a wait list (of threads that are waiting for a find operation to complete at the disk layer)

val mt_initial_state : initial_lim_state:('k, 'v, 'lru) Tjr_lru_cache__.Im_intf.lim_state -> compare_k:('k -> 'k -> int) -> ('k, 'v, 'lru, ('k, ('v, 't) Threading_types.blocked_thread list, unit) Tjr_lib.Tjr_map.map, 't) mt_state

type ('mt_state, 't) with_lru_ops = {

with_lru : a. (lru:'mt_state -> set_lru:('mt_state -> (unit, 't) Tjr_monad.m) -> ('a, 't) Tjr_monad.m) -> ('a, 't) Tjr_monad.m;

}