Module Mt_intf.Threading_types

type 't async = (unit -> (unit, 't) Tjr_monad.m) -> (unit, 't) Tjr_monad.m

The async operation completes with unit almost immediately; the argument may be a long-running computation; it is scheduled for execution.

Because even creating an 'a m in lwt schedules computation, we shield the argument to async to avoid computation.

type ('v, 't) blocked_thread = 'v option -> (unit, 't) Tjr_monad.m

We want to store "blocked threads" in a map, indexed by key; we use a polymap; the threads are of type 'v -> ('a,'t) m; the 'a return type can just be unit, since a thread can launch some other thread to eventually return 'a