Module Tjr_lru_cache__Im_intf

Main lru-in-mem interfaces; don't open (clashing fields)

Cache entries

module Lru_entry : sig ... end

include Lru_entry

type 'v lru_entry =

| Insert of {

value : 'v;

dirty : bool;

}

| Delete of {

dirty : bool;

}

| Lower of 'v option

Cache map entries; values in the map are tagged with a last-accessed time (via underlying Lru impl) and a dirty flag

Entries in the cache for key k:

(Insert v (dirty=true/false)), this occurs on insert

(Delete (dirty=true/false))

(Lower vopt), this occurs when we check the lower layer for a non-existing entry in cache; if we find a value, we insert Lower (Some v), else Lower None; in either case, there is no need to do anything further (ie the entry is not dirty)

(No entry), for a key that hasn't been seen before

val is_Lower : 'a lru_entry -> bool

val mark_clean : 'a lru_entry -> 'a lru_entry

val entry_is_dirty : 'a lru_entry -> bool

val dirty_entry_to_kvop : 'a -> 'b lru_entry -> ('a, 'b) Tjr_fs_shared.Kvop.kvop

Lru cache state (in memory)

module Lru_state_im : sig ... end

The LRU in-memory cache state consists of:

include Lru_state_im

type 'lru lru_state_im = {

max_size : int;

evict_count : int;

lru_state : 'lru;

compare_lru : 'lru -> 'lru -> int;

}

module Lru_ops_im : sig ... end

include Lru_ops_im

type ('k, 'v, 'lru) evictees_x_state = {

evictees : ('k, 'v) Tjr_fs_shared.kvop list;

lru_state_im : 'lru lru_state_im;

}

type ('k, 'v, 'lru) apply_update_from_lower = vopt_from_lower:'v option -> lru_state_im:'lru lru_state_im -> 'v option * ('k, 'v, 'lru) evictees_x_state

type ('k, 'v, 'lru) find_result =

| In_cache of 'v lru_entry

| Not_in_cache
Not_in_cache: we have to call out to disk; what we get back is a 'v option, and we have to then update the lru state (to, in turn, get a possibly different 'v option and the evictees_x_state; actually, since we are the only thread operating on 'k (other threads block), the 'v option returned is that received from lower)

type ('k, 'v, 'lru) lru_ops_im = {

find : 'k -> 'lru lru_state_im -> [ `In of 'v option | `Not_in ];

insert : 'k -> 'v -> 'lru lru_state_im -> ('k, 'v, 'lru) evictees_x_state;

update_from_lower : 'k -> 'v option -> 'lru lru_state_im -> ('k, 'v, 'lru) evictees_x_state;

delete : 'k -> 'lru lru_state_im -> ('k, 'v, 'lru) evictees_x_state;

sync_key : 'k -> 'lru lru_state_im -> (('k, 'v) Tjr_fs_shared.kvop * 'lru lru_state_im) option;
If the key is present and the entry is dirty, we get a kvop and a state with that entry cleaned; otherwise we get None. NOTE in the multithreaded code, this must also call out to the lower layer. The state is changed if the entry is dirty.

sync_all_keys : 'lru lru_state_im -> ('k, 'v, 'lru) evictees_x_state;
NOTE in multithreaded code, this must call lower before returning

}

This type is what is returned by the make_lru_in_mem function.

NOTE that sync_key performs only the in-mem updates (ie clearing the dirty flag). If you want to flush to disk, you have to do something else (see Lru_multithreaded).

NOTE find returns a "find_result", which consists of a callback function which tries to retrieve from below, and then on return updates the then-current lim_state (which may already have the entry...); this is implemented in the multithreaded code

NOTE the interface tries to favour pure state passing; find is the only routine which may need to call to disk

type lru_params_im = < max_size : int; evict_count : int; >

type ('k, 'v, 'lru) lru_factory_im = < empty : lru_params_im -> 'lru lru_state_im; lru_ops_im : ('k, 'v, 'lru) lru_ops_im; >