Module Tjr_lru_cache__Im_intf.Lru_ops_im

Lru ops (in memory)

type ('k, 'v, 'lru) evictees_x_state = {

evictees : ('k, 'v) Tjr_fs_shared.kvop list;

lru_state_im : 'lru lru_state_im;

}

type ('k, 'v, 'lru) apply_update_from_lower = vopt_from_lower:'v option -> lru_state_im:'lru lru_state_im -> 'v option * ('k, 'v, 'lru) evictees_x_state

type ('k, 'v, 'lru) find_result =

| In_cache of 'v lru_entry

| Not_in_cache
Not_in_cache: we have to call out to disk; what we get back is a 'v option, and we have to then update the lru state (to, in turn, get a possibly different 'v option and the evictees_x_state; actually, since we are the only thread operating on 'k (other threads block), the 'v option returned is that received from lower)

type ('k, 'v, 'lru) lru_ops_im = {

find : 'k -> 'lru lru_state_im -> [ `In of 'v option | `Not_in ];

insert : 'k -> 'v -> 'lru lru_state_im -> ('k, 'v, 'lru) evictees_x_state;

update_from_lower : 'k -> 'v option -> 'lru lru_state_im -> ('k, 'v, 'lru) evictees_x_state;

delete : 'k -> 'lru lru_state_im -> ('k, 'v, 'lru) evictees_x_state;

sync_key : 'k -> 'lru lru_state_im -> (('k, 'v) Tjr_fs_shared.kvop * 'lru lru_state_im) option;
If the key is present and the entry is dirty, we get a kvop and a state with that entry cleaned; otherwise we get None. NOTE in the multithreaded code, this must also call out to the lower layer. The state is changed if the entry is dirty.

sync_all_keys : 'lru lru_state_im -> ('k, 'v, 'lru) evictees_x_state;
NOTE in multithreaded code, this must call lower before returning

}

This type is what is returned by the make_lru_in_mem function.

NOTE that sync_key performs only the in-mem updates (ie clearing the dirty flag). If you want to flush to disk, you have to do something else (see Lru_multithreaded).

NOTE find returns a "find_result", which consists of a callback function which tries to retrieve from below, and then on return updates the then-current lim_state (which may already have the entry...); this is implemented in the multithreaded code

NOTE the interface tries to favour pure state passing; find is the only routine which may need to call to disk