Module Tjr_lru_cache__Lru_in_mem

A single-threaded in-memory LRU cache, as a prelude to the real thing.

We assume all operations complete quickly from the cache. Operations that need to go to disk are modelled using explicit continuation passing.

The difficulty is how to extend this to deal with disk-backed eviction.

NOTE that the operations occur not in a monad - instead, explicit state passing is used, and we return the possible evictees; this is a prelude to the real version where we need to do something with the evictees

module type S = sig ... end

module type T = sig ... end

module Make_v1 : functor (S : S) -> sig ... end

With full sig

module Make_v2 : functor (S : S) -> T with module S = S

module Make = Make_v2

module Examples : sig ... end