Module Tjr_lru_cache__Mt_intf.Mt_state

type ('k, 'v, 'lru, 't_map, 't) mt_state = {

lru_state_im : 'lru Tjr_lru_cache.Im_intf.lru_state_im;

blocked_threads : 't_map;

blocked_threads_ops : ('k, ('v, 't) blocked_thread list, 't_map) Tjr_lib.Tjr_map.map_ops;

}

The state consists of:

the cache state

a map from k to a wait list (of threads that are waiting for a find operation to complete at the disk layer)

NOTE t_map is "thread map"

val mt_initial_state : lru_state_im:'lru Tjr_lru_cache.Im_intf.lru_state_im -> k_cmp:('k -> 'k -> int) -> ('k, 'v, 'lru, ('k, ('v, 't) blocked_thread list, unit) Tjr_lib.Tjr_map.map, 't) mt_state