Module Tjr_lru_cache__Multithreaded_lru

Non-functorial multithreaded / concurrent-safe implementation of the LRU interface. This extends the in-memory explicit-state-passing code to the monad.

Compared to the in-mem version:

we use the monad

we allow multiple threads to call the interface methods

for long-running operations (eg find, when the key is not in the i-m cache), we maintain a queue of blocked threads, make a single call to the lower layers, and when this returns we release the threads

we have explicit messages to the lower layer

the api includes "persist mode" (now or later) to indicate whether insert/delete operations should persist immediately or not

Given that the likely target is Lwt, this is probably overkill.

Auxiliary defns

module Internal2 : sig ... end

val make_lru_callback_ops : lim_ops:('k, 'v, 'lru) Tjr_lru_cache.Im_intf.lim_ops -> monad_ops:'t Tjr_monad.monad_ops -> async:((unit -> (unit, 't) Tjr_monad.m) -> (unit, 't) Tjr_monad.m) -> with_lru:(('k, 'v, 'lru, 'a, 't) Tjr_lru_cache.Mt_intf.Mt_state_type.mt_state, 't) Tjr_monad.with_state -> to_lower:(('k, 'v, 't) Tjr_lru_cache.Mt_intf.Msg_type.msg -> (unit, 't) Tjr_monad.m) -> ('k, 'v, 't) Tjr_lru_cache.Mt_intf.Mt_callback_ops.mt_callback_ops

val make_lru_ops : monad_ops:'t Tjr_monad.monad_ops -> event_ops:'t Tjr_monad.Event.event_ops -> callback_ops:('k, 'v, 't) Tjr_lru_cache.Mt_intf.Mt_callback_ops.mt_callback_ops -> ('k, 'v, 't) Tjr_lru_cache.Mt_intf.Mt_ops.mt_ops

Construct the multithreaded LRU

val make_multithreaded_lru : lru_fc:('k, 'v, 'lru) Tjr_lru_cache.Im_intf.lru_fc -> monad_ops:'t Tjr_monad.monad_ops -> async:((unit -> (unit, 't) Tjr_monad.m) -> (unit, 't) Tjr_monad.m) -> event_ops:'t Tjr_monad.Event.event_ops -> with_lru:(('k, 'v, 'lru, 'a, 't) Tjr_lru_cache.Mt_intf.Mt_state_type.mt_state, 't) Tjr_monad.with_state -> to_lower:(('k, 'v, 't) Tjr_lru_cache.Mt_intf.Msg_type.msg -> (unit, 't) Tjr_monad.m) -> ('k, 'v, 't) Tjr_lru_cache.Mt_intf.Mt_ops.mt_ops

NOTE see also Lru_in_mem.Make_lru_fc; FIXME should also return an initial state to avoid remembering how to to construct this later