Module Tjr_lru_cache__Multithreaded_lru.Internal2

val add_callback_on_key : callback:('a, 'b) Tjr_lru_cache__Mt_intf.Threading_types.blocked_thread -> k:'c -> lru:('c, 'a, 'd, 'e, 'b) Tjr_lru_cache.Mt_intf.Mt_state_type.mt_state -> ('c, 'a, 'd, 'e, 'b) Tjr_lru_cache.Mt_intf.Mt_state_type.mt_state * [> `Lower_call_in_progress | `Need_to_call_lower ]

NOTE the following, which essentially adds waiting threads to a wait queue, and notifies them when an event occurs, is a re-implementation of a concurrency staple (eg Lwt tasks); the implementation here is explicit because we may want to verify the implementation using the local state blocked_threads (rather than some generic version provided by the monad)

val run_callbacks_for_key : monad_ops:'t Tjr_monad.monad_ops -> async:'t Tjr_lru_cache.Mt_intf.Threading_types.async -> k:'a -> vopt_from_lower:'b option -> lru:('a, 'b, 'c, 'd, 't) Tjr_lru_cache.Mt_intf.Mt_state_type.mt_state -> (('a, 'b, 'c, 'd, 't) Tjr_lru_cache.Mt_intf.Mt_state_type.mt_state, 't) Tjr_monad.m

NOTE this also removes all waiters on key k

Callback interface

val make_lru_callback_ops : lim_ops:('k, 'v, 'lru) Tjr_lru_cache__.Im_intf.lim_ops -> monad_ops:'t Tjr_monad.monad_ops -> async:'t Tjr_lru_cache.Mt_intf.Threading_types.async -> with_lru:(('k, 'v, 'lru, 'a, 't) Tjr_lru_cache.Mt_intf.Mt_state_type.mt_state, 't) Tjr_lru_cache.Mt_intf.Mt_state_type.with_lru_ops -> to_lower:(('k, 'v, 't) Tjr_lru_cache.Mt_intf.Msg_type.msg -> (unit, 't) Tjr_monad.m) -> ('k, 'v, 't) Tjr_lru_cache.Mt_intf.Mt_callback_ops.mt_callback_ops

NOTE with_lru f executes whilst locking lru state; it should be possible to execute the monadic f quickly, without yielding to other threads.

For find:

We access the lru_state (exlusively/serially).

We check if the find would succeed without going to disk. If so, we are done.

Otherwise, we check the wait list for the given key. If this is non-existent, we create it, execute the lower call to find, and on return we alter the then-current cache state (updates are serialized), and notify all pending threads on the wait list, and delete the wait list.

For Persist_later insert and delete, we access the lru_state, modify the cache state. If we have evictees, we need to flush these to lower. The problem is that we may not want to hold the lru_state during this eviction. At the moment, we hold the lock.

For Persist_now insert and delete, we do as above. If we have evictees, we can add the op to the evictees list (so, the size of the evictees should be 1 less than the amount that can safely fit in one block FIXME) and then call the lower layer.

For sync_key (and sync_keys) we simply take the entry in the cache (if any) and mark it clean and flush to lower.

Event-based interface

We now implement the standard (non-callback) interface, using events.

type 't event_ops = 't Tjr_monad.Event.event_ops

We assume that there is a way to "fulfill" an 'a m with an 'a

val make_lru_ops : monad_ops:'t Tjr_monad.monad_ops -> event_ops:'t Tjr_monad.Event.event_ops -> callback_ops:('k, 'v, 't) Tjr_lru_cache.Mt_intf.Mt_callback_ops.mt_callback_ops -> ('k, 'v, 't) Tjr_lru_cache.Mt_intf.Mt_ops.mt_ops