Module Tjr_lru_cache__Multithreaded_lru.Make

Parameters

S : S

Signature

module S = S

val (>>=) : ('a, S.t) Tjr_monad.m -> ('a -> ('b, S.t) Tjr_monad.m) -> ('b, S.t) Tjr_monad.m

val return : 'a -> ('a, S.t) Tjr_monad.m

Util

val add_callback_on_key : callback:('a, 'b) Tjr_lru_cache.Mt_intf.blocked_thread -> k:'c -> lru:('c, 'a, 'd, 'e, 'b) Tjr_lru_cache.Mt_intf.mt_state -> ('c, 'a, 'd, 'e, 'b) Tjr_lru_cache.Mt_intf.mt_state * [> `Lower_call_in_progress | `Need_to_call_lower ]

NOTE the following, which essentially adds waiting threads to a wait queue, and notifies them when an event occurs, is a re-implementation of a concurrency staple (eg Lwt tasks); the implementation here is explicit because we may want to verify the implementation using the local state blocked_threads (rather than some generic version provided by the monad)

NOTE there is no entry in the blocked_threads iff there is no lower call in progress

val run_callbacks_for_key : k:'a -> vopt_from_lower:'b option -> lru:('a, 'b, 'c, 'd, S.t) Tjr_lru_cache.Mt_intf.mt_state -> (('a, 'b, 'c, 'd, S.t) Tjr_lru_cache.Mt_intf.mt_state, S.t) Tjr_monad.m

NOTE this also removes all waiters on key k

Callback interface

val make_callback_ops : lru_ops_im:('k, 'v, 'lru) Tjr_lru_cache.Im_intf.lru_ops_im -> with_lru:(('k, 'v, 'lru, 'a, S.t) Tjr_lru_cache.Mt_intf.mt_state, S.t) Tjr_monad.with_state -> to_lower:(('k, 'v, S.t) Tjr_lru_cache.Mt_intf.lru_msg -> (unit, S.t) Tjr_monad.m) -> ('k, 'v, S.t) Tjr_lru_cache.Mt_intf.Mt_callback_ops.mt_callback_ops

NOTE with_lru f executes whilst locking lru state; it should be possible to execute the monadic f quickly, without yielding to other threads.

For find:

We access the lru_state (exlusively/serially).

We check if the find would succeed without going to disk. If so, we are done.

Otherwise, we check the wait list for the given key. If this is non-existent, we create it, execute the lower call to find, and on return we alter the then-current cache state (updates are serialized), and notify all pending threads on the wait list, and delete the wait list.

For Persist_later insert and delete, we access the lru_state, modify the cache state. If we have evictees, we need to flush these to lower. The problem is that we may not want to hold the lru_state during this eviction. At the moment, we hold the lock.

For Persist_now insert and delete, we do as above. If we have evictees, we can add the op to the evictees list (so, the size of the evictees should be 1 less than the amount that can safely fit in one block FIXME) and then call the lower layer.

For sync_key (and sync_keys) we simply take the entry in the cache (if any) and mark it clean and flush to lower.

Event-based interface

We now implement the standard (non-callback) interface, using events.

val make_lru_ops : callback_ops:('k, 'v, S.t) Tjr_lru_cache.Mt_intf.Mt_callback_ops.mt_callback_ops -> ('k, 'v, S.t) Tjr_lru_cache.Mt_intf.Mt_ops.mt_ops

Join event-based interface to callback interface

val make : lru_ops_im:('a, 'b, 'c) Tjr_lru_cache.Im_intf.lru_ops_im -> with_lru:(('a, 'b, 'c, 'd, S.t) Tjr_lru_cache.Mt_intf.mt_state, S.t) Tjr_monad.with_state -> to_lower:(('a, 'b, S.t) Tjr_lru_cache.Mt_intf.lru_msg -> (unit, S.t) Tjr_monad.m) -> ('a, 'b, S.t) Tjr_lru_cache.Mt_intf.Mt_ops.mt_ops