Module Tjr_lru_cache

A multithreaded LRU cache (part of ImpFS; don't open)

NOTE probably best not to open this module, because msg_type and entry have constructors like Insert etc that clash with others defined elsewhere

module $Open__44_864 = Pvt_in_mem

module $Open__45_880 = Pvt_in_mem.Lru_in_mem

Multi-threaded version

Interfaces

module Mt_intf : sig ... end

Main mt (multithreaded lru) types.

Persistence mode

include Mt_intf.Persist_mode

type persist_mode =

| Persist_later

| Persist_now

A calling mode can be "later" (ie, perform immediately in mem and return; persist sometime later) or "now", in which case we can optionally supply a callback. A call can be blocking or non-blocking. For non-blocking, the calling thread is expected to launch an async promise that resolves without blocking the caller. Thus, we implement only blocking versions of calls.

The resulting multithreaded operations

include Mt_intf.Mt_ops

type ('k, 'v, 't) mt_ops = {

mt_find : 'k -> ('v option, 't) Tjr_monad.m;

mt_insert : Mt_intf.persist_mode -> 'k -> 'v -> (unit, 't) Tjr_monad.m;

mt_delete : Mt_intf.persist_mode -> 'k -> (unit, 't) Tjr_monad.m;

mt_sync_key : 'k -> (unit, 't) Tjr_monad.m;

mt_sync_all_keys : unit -> (unit, 't) Tjr_monad.m;

}

Construct the multithreaded LRU (functor version)

module type S = sig ... end

module type T = sig ... end

module Make : functor (S : S) -> T with type k = S.k and type v = S.v and type t = S.t

NOTE we expect the result of this Make to be bound to a module "Lru" or "Lru_"