Module Tjr_lru_cache

A multithreaded LRU cache (part of ImpFS; don't open)

NOTE probably best not to open this module, because msg_type and entry have constructors like Insert etc that clash with others defined elsewhere

Single-threaded version

module Im_intf : sig ... end

Main lru-in-mem interfaces; usually don't open this

module Lru_in_mem : sig ... end

A single-threaded in-memory LRU cache, as a prelude to the real thing.

Multi-threaded version

Interfaces

module Mt_intf : sig ... end

Main mt (multithreaded lru) types.

Persistence mode

The resulting multithreaded operations

include Mt_intf.Mt_ops

type ('k, 'v, 't) mt_ops = {

mt_find : 'k -> ('v option, 't) Tjr_monad.m;

mt_insert : Mt_intf.persist_mode -> 'k -> 'v -> (unit, 't) Tjr_monad.m;

mt_delete : Mt_intf.persist_mode -> 'k -> (unit, 't) Tjr_monad.m;

mt_sync_key : 'k -> (unit, 't) Tjr_monad.m;

mt_sync_all_keys : unit -> (unit, 't) Tjr_monad.m;

}

Non-functorial mt impl

module Multithreaded_lru : sig ... end

Non-functorial multithreaded / concurrent-safe implementation of the LRU interface. This extends the in-memory explicit-state-passing code to the monad.

Construct the multithreaded LRU (functor version)

module Make : sig ... end

Performance testing

module Test_performance : sig ... end

Performance testing code.